% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_2048.R
\name{play_2048}
\alias{play_2048}
\alias{2048}
\alias{twenty48}
\alias{resume_2048}
\title{Play 2048 in the console}
\usage{
play_2048(size = 4, dynamic = rstudioapi::isAvailable())

resume_2048()
}
\arguments{
\item{size}{An integer giving the grid's dimensions.
Defaults to \code{4}, which generates a 4-by-4 grid.}

\item{dynamic}{If \code{dynamic} is \code{TRUE}, commands can be input without pressing
enter, creating a more seamless game experience.
Note that \code{dynamic} commands currently only work in RStudio.}
}
\value{
Generates an interactive game of 2048 in the console.
}
\description{
Interact with the game by typing commands into the console and
pressing enter.
}
\examples{
play_2048()
play_2048(size = 5)
}
