% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twenty48-package.R
\docType{package}
\name{twenty48-package}
\alias{twenty48-package}
\alias{_PACKAGE}
\title{twenty48: Play a Game of 2048 in the Console}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Generates a game of 2048 that can be played in the console. Supports grids of arbitrary sizes, undoing the last move, and resuming a game that was exited during the current session.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rossellhayes/twenty48}
  \item Report bugs at \url{https://github.com/rossellhayes/twenty48/issues}
}

}
\author{
\strong{Maintainer}: Alexander Rossell Hayes \email{alexander@rossellhayes.com} (\href{https://orcid.org/0000-0001-9412-0457}{ORCID}) [copyright holder]

}
\keyword{internal}
