% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_2048.R
\name{play_2048}
\alias{play_2048}
\alias{resume_2048}
\title{Play 2048 in the console}
\usage{
play_2048(size = 4)

resume_2048()
}
\arguments{
\item{size}{A vector of one or two integers giving the grid's dimensions.
If length one, a square grid is generated.
Defaults to \code{4}, which generates a 4-by-4 grid.}
}
\value{
Generates an interactive game of 2048 in the console.
}
\description{
Play 2048 in the console
}
