% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_2048.R
\name{play_2048}
\alias{play_2048}
\alias{resume_2048}
\title{Play 2048 in the console}
\usage{
play_2048(size = 4)

resume_2048()
}
\arguments{
\item{size}{An integer giving the grid's dimensions.
Defaults to \code{4}, which generates a 4-by-4 grid.}
}
\value{
Generates an interactive game of 2048 in the console.
}
\description{
Interact with the game by typing commands into the console and
pressing enter.
}
\examples{
play_2048()

resume_2048()
}
